streamlit_app/requirements.txt
streamlit
pillow
torch
opencv-python
numpy
